Implement authorization middleware in some route

+++ app/Http/Controllers/ManageCategoryController.php
@@ -7,12 +7,20 @@ use Illuminate\Http\Request;
 
 class ManageCategoryController extends Controller {
     public function index() {
+
+        // Authorization
+        $this->authorize('admin');
+
         return view('manage-category', [
             'categories' => Category::paginate(10)
         ]);
     }
 
     public function store(Request $request) {
+
+        // Authorization
+        $this->authorize('admin');
+
         $request->validate([
             'category' => ['required', 'unique:categories,name'],
         ]);
@@ -25,6 +33,10 @@ class ManageCategoryController extends Controller {
     }
 
     public function update(Category $category, Request $request) {
+
+        // Authorization
+        $this->authorize('admin');
+
         $request->validate([
             'category' => ['required', 'unique:categories,name'],
         ]);
@@ -39,7 +51,9 @@ class ManageCategoryController extends Controller {
     }
 
     public function delete(Category $category) {
-        //AUTHORIZATION
+
+        // Authorization
+        $this->authorize('admin');
 
         $categoryName = $category->name;
         $category->games()->delete();
